Add FileIngestionEngine
This creates a FileIngestionEngine. It is a concrete
implementation of AbstractIngestionEngine, which reads in a file,
converts it into an MPEG-TS stream represented as a series of Payloads,
and then pipes that to Kafka. This mostly defines the input stream
as a file stream.

Also adds an integration test for file ingestion and fixes some
errors in the AbstractIngestionEngine uncovered by this test.